+++ ssl/ssl_local.h
@@ -1216,6 +1216,8 @@ struct ssl_ctx_st {
 # ifndef OPENSSL_NO_QLOG
     char *qlog_title; /* Session title for qlog */
 # endif
+
+    struct ossl_grease_st *grease;
 };
 
 typedef struct ossl_quic_tls_callbacks_st {
@@ -3170,5 +3172,18 @@ long ossl_ctrl_internal(SSL *s, int cmd, long larg, void *parg, int no_quic);
      SSL_DOMAIN_FLAG_THREAD_ASSISTED |          \
      SSL_DOMAIN_FLAG_BLOCKING |                 \
      SSL_DOMAIN_FLAG_LEGACY_BLOCKING)
+
+/* EXTENSION < 0xFFFF is the actual extension value */
+# define OSSL_GREASE_CIPHER 0x10001
+
+struct ossl_grease_st {
+    struct ossl_grease_st *next;
+    unsigned int extension;
+    int new_extension;
+    unsigned int value;
+    unsigned char *data;
+    size_t length;
+};
+typedef struct ossl_grease_st OSSL_GREASE;
 
 #endif